+++ test/scum/scum_ui_test.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:math';
 
 import 'package:cards_with_cats/cards/card.dart';